move info module paths to separate file
