+++ src/app/info/info-routing-paths.ts
@@ -1,0 +1,16 @@
+import { getInfoModulePath } from '../app-routing-paths';
+
+export const END_USER_AGREEMENT_PATH = 'end-user-agreement';
+export const PRIVACY_PATH = 'privacy';
+
+export function getEndUserAgreementPath() {
+    return getSubPath(END_USER_AGREEMENT_PATH);
+}
+
+export function getPrivacyPath() {
+    return getSubPath(PRIVACY_PATH);
+}
+
+function getSubPath(path: string) {
+    return `${getInfoModulePath()}/${path}`;
+}
