+++ src/app/info/info-routing.module.ts
@@ -1,24 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { EndUserAgreementComponent } from './end-user-agreement/end-user-agreement.component';
-import { getInfoModulePath } from '../app-routing-paths';
 import { I18nBreadcrumbResolver } from '../core/breadcrumbs/i18n-breadcrumb.resolver';
 import { PrivacyComponent } from './privacy/privacy.component';
-
-const END_USER_AGREEMENT_PATH = 'end-user-agreement';
-const PRIVACY_PATH = 'privacy';
-
-export function getEndUserAgreementPath() {
-  return getSubPath(END_USER_AGREEMENT_PATH);
-}
-
-export function getPrivacyPath() {
-  return getSubPath(PRIVACY_PATH);
-}
-
-function getSubPath(path: string) {
-  return `${getInfoModulePath()}/${path}`;
-}
+import { PRIVACY_PATH, END_USER_AGREEMENT_PATH } from './info-routing-paths';
 
 @NgModule({
   imports: [
